Move documentutls, spreadsheet and tempfiles to pkg
