+++ pkg/documentutils/documentbase/DocumentBase.go
@@ -1,8 +1,8 @@
 package documentbase
 
 import (
+	"github.com/asciich/asciichgolangpublic/pkg/spreadsheet"
 	"github.com/asciich/asciichgolangpublic/pkg/tracederrors"
-	"github.com/asciich/asciichgolangpublic/spreadsheet"
 )
 
 type DocumentBase struct {
