+++ pkg/documentutils/documentbase/Text.go
