+++ pkg/documentutils/htmldocument/HtmlDocument.go
@@ -1,6 +1,6 @@
 package htmldocument
 
-import "github.com/asciich/asciichgolangpublic/documentutils/documentbase"
+import "github.com/asciich/asciichgolangpublic/pkg/documentutils/documentbase"
 
 type HtmlDocument struct {
 	documentbase.DocumentBase
