+++ pkg/tmuxutils/TmuxWindow_test.go
@@ -8,9 +8,9 @@ import (
 	"github.com/asciich/asciichgolangpublic/files"
 	"github.com/asciich/asciichgolangpublic/parameteroptions"
 	"github.com/asciich/asciichgolangpublic/pkg/contextutils"
-	"github.com/asciich/asciichgolangpublic/pkg/tmuxutils"
-	"github.com/asciich/asciichgolangpublic/tempfiles"
+	"github.com/asciich/asciichgolangpublic/pkg/filesutils/tempfiles"
 	"github.com/asciich/asciichgolangpublic/pkg/testutils"
+	"github.com/asciich/asciichgolangpublic/pkg/tmuxutils"
 )
 
 func TestTemuxWindow_CreateAndDeleteWindow(t *testing.T) {
